feat: use new performance apis for search campaigns
Use the updated performance apis that do server-based aggregation, so the screens do not overwhelm the browser on large campaigns.

To introduce this incrementally, this is being added for search campaigns only initially, with the intention of subsequently switching other campaign types to use this in due course.

Although the original intention was to keep the UI the same, this required a lot of throw-away work, so I've brought in the experience of what we now do on our internal dashboards.

Outstanding steps:
- [ ] re-instate date filtering
- [ ] ad / adset breakdowns
- [ ] review list of metrics