refactor: switch to using `@/` for in-app imports
Use `@/name` for app imports, as opposed to just `name` as previously. This is the modern convention for web apps, and makes clearer when imports are coming from the application as opposed to npm modules.

There's a few deep `../../` style imports remaining - these get worse if I move them to `@/`-style imports, and I think are better left until we move the files around a bit more in a subsequent PR.