fix: don't refresh search preview data
The default of swr is to refresh the data on screen activation, network reconnect etc. This is not needed for the search preview data as the data is very unlikely to change within one user session, generating needless requests that return the same data.

See https://swr.vercel.app/docs/revalidation#disable-automatic-revalidations