fix: react-router was breaking
Overzealous upgrading of overrides by renovate means that path-to-regexp got upragded from 1.9 to 8.2, which breaks react-router@5. On staging, ads-ui simply doesn't load as a result.

Switch the overrides section back to how it used to be, explicilty describing the vulnerable versions to avoid, rather than randomly picking the latest version.