fix: react-router was breaking (#1444)
Overzealous upgrading of overrides by renovate means that path-to-regexp
got upgraded from 1.9 to 8.2, which breaks react-router@5. On staging,
ads-ui simply doesn't load as a result.

Switch the overrides section back to how it used to be, explicitly
describing the vulnerable versions to avoid, rather than randomly
picking the latest version. Then renovate can actually pick the correct
version to use.