chore(renovate): adjust schedule (#1470)
It's very distracting to get renovate PRs during the working day, as we
try to keep to "PR zero" to reduce the amount of wasted effort not yet
delivering value. A constant stream of renovate PRs throughout the day
also encourages ignoring them, a poor outcome.

Also, bring across from our other repos the clearer phrasing of the
config to prevent updating of the override ranges.