chore: actually enable eslint (#1611)
It appears that eslint wasn't correctly running against the codebase, I
think as a result of an older failed upgrade to eslint 9.

Make it apply with our standard settings, and either fix or suppress the
98 errors that were present.

Due to the quantity of errors, I've had to take the approach of just
suppressing many of them so at least we can ensure we do not add new
errors going forward. We should tackle all the eslint ignores over time.


Draft because we need to do a fair bit of testing and review that the
changes I've made here haven't broken anything.